jpa rep
todo: change profile into qualifier

+++ src/main/java/com/springmvc/services/UserService.java
@@ -7,6 +7,6 @@ import java.util.List;
 public interface UserService {
     void addUser(User user);
     List<User> getAllUsers();
-    User getUserById(int id);
-    void deleteUserById(int id);
+    User getUserById(long id);
+    void deleteUserById(long id);
 }
