+++ src/main/java/com/springmvc/services/impl/UserServiceImpl.java
@@ -24,12 +24,12 @@ public class UserServiceImpl implements UserService {
     }
 
     @Override
-    public User getUserById(int id) {
+    public User getUserById(long id) {
         return userRepository.findById(id);
     }
 
     @Override
-    public void deleteUserById(int id) {
+    public void deleteUserById(long id) {
         userRepository.deleteById(id);
     }
 }
